+++ src/App.vue
@@ -1,8 +1,9 @@
 <template>
   <div id="app">
-    <Navigation id="app_navigation"></Navigation>
-    <div class="clearfix"></div>
-    <router-view id="app_main"></router-view>
+    <router-view />
+    <!-- <Navigation id="app_navigation"></Navigation> -->
+    <!-- <div class="clearfix"></div> -->
+    <!-- <router-view id="app_main"></router-view> -->
   </div>
 </template>
 
@@ -33,17 +34,46 @@ export default {
     -moz-osx-font-smoothing: grayscale;
     height: 100%;
     max-height: 100%;
-    #app_navigation{
-      height: 61px!important;
+  }
+  .HomeIndex, .DatabaseIndex{
+    h1{
+      font-size: 30px;
+      padding: 100px;
+      margin: 0;
+      text-align: center;
       background: #4A524F;
+      color: gold;
     }
-    #app_main{
-      height: calc(100% - 61px) !important;
-      max-height: calc(100% - 61px) !important;
-      overflow-y: auto;
-    }
-    .clearfix{
-      clear:both
+    .wrapper{
+      max-width: 1000px;
+      margin: auto;
+      a{
+        font-weight: bolder;
+        font-size: larger;
+        text-align: center!important;
+        &:hover{
+          .el-card{
+            background: darkslategray!important;
+            color: #ffffff;
+          }
+        }
+      }
+      .el-row{
+        // margin: 50px 0;
+      }
+      .el-card__body{
+        padding: 50px;
+        text-align: center!important;
+        i{
+          font-size: x-large;
+        }
+      }
+      .el-card{
+        margin-bottom: 10px!important;
+      }
+      .add .el-card__body{
+        padding: 30px!important;
+      }
     }
   }
 </style>
