+++ src/components/database/widgets/Footer.vue
