+++ src/components/database/widgets/Navigation.vue
