+++ src/components/home/HomeDatabase.vue
@@ -1,0 +1,3 @@
+<template>
+  <div class="HomeDataBase"></div>
+</template>
