+++ src/components/home/HomeIndex.vue
@@ -1,46 +1,23 @@
 <template>
   <div class="HomeIndex">
+    <h1>
+      欢迎来到，基础制造工艺资源环境负荷数据库
+    </h1>
     <div class="wrapper">
-      <h1>
-        欢迎来到，基础制造工艺资源环境负荷数据库
-      </h1>
-      <el-row :gutter="12">
-        <el-col :span="8">
-          <router-link :to="{name: 'SceneIndex'}">
+      <el-row :gutter="20" v-for="sub in navList" :key="sub.index">
+        <el-col :span="12" v-for="item in sub.children" :key="item.index">
+          <router-link :to="{name: 'DatabaseIndex', params:{categoryId: item['id']}}">
             <el-card always="always">
-              工艺场景管理
+              <i class="fa fa-fw fa-database"></i>  {{ item['title'] }}数据库
             </el-card>
           </router-link>
         </el-col>
-        <el-col :span="8">
-          <router-link :to="{name: 'ManageIndex'}">
-            <el-card always="always">
-              基础数据管理
-            </el-card>
-          </router-link>
-        </el-col>
-        <el-col :span="8">
-          <router-link :to="{name: 'BatchIndex'}">
-            <el-card always="always">
-              批量数据处理
-            </el-card>
-          </router-link>
-        </el-col>
-      </el-row>
-      <el-row :gutter="12">
-        <el-col :span="12">
-          <router-link :to="{name: 'SearchIndex'}">
-            <el-card always="always">
-              <i class="fa fa-fw fa-search"></i> RESTful 查询
-            </el-card>
-          </router-link>
-        </el-col>
-        <el-col :span="12">
-          <router-link :to="{name: 'UserIndex'}">
-            <el-card always="always">
-              <i class="fa fa-fw fa-users"></i> 用户权限管理
-            </el-card>
-          </router-link>
+        <el-col :span="24">
+          <el-card class="add">
+            <router-link :to="{name: 'HomeIndex'}">
+              <i class="fa fa-fw fa-plus-circle"></i> 添加数据库
+            </router-link>
+          </el-card>
         </el-col>
       </el-row>
     </div>
@@ -50,9 +27,22 @@
 <script>
 export default {
   name: 'HomeIndex',
-  methods: {}
+  computed: {
+    navList () {
+      return this.$store.state.categories
+    }
+  },
+  data () {
+    return {
+    }
+  }
 }
 </script>
 
 <style lang="scss">
+.HomeIndex{
+  h1{
+    margin-bottom: 50px;
+  }
+}
 </style>
